updated package-info.java moved PersonAddress to facade layer where it belongs
