+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/facade/dto/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Contains Domain Transfer Objects (DTO) or composite objects. Used to make interaction with one or more service easier
+ * when associations may not normally exist.
+ */
+package ca.mikegabelmann.demo2.persistence.facade.dto;
