+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/facade/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains facade layer which handles interactions between 1 or more services.
+ */
+package ca.mikegabelmann.demo2.persistence.facade;
