+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/model/package-info.java
@@ -1,6 +1,4 @@
 /**
  * This package contains all the classes that support mapping between the database and application using JPA.
- *
- * @author mgabelmann
  */
 package ca.mikegabelmann.demo2.persistence.model;
