+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/repository/package-info.java
@@ -1,6 +1,4 @@
 /**
  * This package contains all the Data Access Objects (DAO) for use by the service/business layer of the application.
- *
- * @author mgabelmann
  */
 package ca.mikegabelmann.demo2.persistence.repository;
